requirements.txt
numpy==1.20.3
pandas==1.3.3
scipy==1.4.1
scikit-learn==1.1.1
GPUtil==1.4.0
torch==1.10.1
PyYAML~=5.1.2
tqdm==4.62.3
matplotlib~=3.4.3
joblib~=1.0.1
seaborn~=0.11.2
